feat: outsource types
